Fix follow() method to prevent a double content request
The follow method gets the response text from line 145 and passes it to a new HAL object in line 146, however it passes the text as a non-named parameter.
As a result, the HAL object constructor ignores this response text, and calls a request to `response.url` _again_ when it constructs the object.
This fix passes the `response.text` to the named `json_str` parameter, as I believe was originally intended. This prevents the constructor calling request on the url again.
I've tested this fix in my project and it doesn't seem to break anything.